+++ .gitignore
@@ -1,4 +1,5 @@
-node_modules/
 .idea/
-build/
 package-lock.json
+/node_modules/
+/public/build/
+.DS_Store
