+++ Migration.md
@@ -6,3 +6,8 @@
   - Hotkeys(target).on(action_name,handler1,handler2)
   - Hotkeys(target).off(action_name,handler1,handler2)
     - are now the preferred way of binding actions
+
+
+### 1.0.0 => 2.0.0    
+- migrating to rollup & svelte 
+  - provide separete builds for core & svelte components
