+++ src/core/mousetrap-wrapper.js
