+++ src/example/example-bindings.js
