+++ src/example/example.js
