+++ src/main.ts
@@ -1,9 +1,11 @@
 import App from './App.svelte';
 
+
+
 const app = new App({
 	target: document.body,
 	props: {
-		name: 'world'
+		name: 'world2'
 	}
 });
